BUG#14309019 - mysqlbinlog -v does not properly decode                DECIMAL values in an RBR log.
=== Description ===

mysqlbinlog -v reads a binlog file and prints it in a
readable format. In case of decimal values, it converts
the number in base of pow(10,9). It used to print a
decimal point after every block of 9 digits. As a result,
decimal point was sometimes printed more than once or
sometimes mispalced. As an example, The value: -0.938582
was printed as -938582000.000000000 and 4294967296.001 as
000000004.294967296.001000000.000000000 

=== Fix ===

Modified the for loop used for printing digits of the
decimal value. Broke it into two for loops. Below is the
sequence of steps followed:
1) A for loop prints digits before the decimal point,
2) A decimal point is printed between the two for loops.
3) The second for loop prints digits after the decimal
   point.