BUG#13875534 - TESTS FOR OUT-OF-BOUND AND DEFAULT VALUES OF VARIABLES IN                SYS_VARS SUITE IN MTR
Description:
-----------
In mysql-test/suite/sys_vars/t/*.test, in each file there should be a test
that tries to assign out-of-bounds values to the variables, and then does a
SHOW to ensure that it is being clamped to the min or max.

Also, there should be a test that tries to assign SET GLOBAL (name_of_var)=
DEFAULT.

Suggested Fix:
-------------
Review the entire sys_vars suite and make the
necessary enhancements.