Bug#14551372 - INNODB: UNIV_SYNC_DEBUG SYNC ARRAY IS NOT OPTIMIZED
As part of the investigation into Bug 14520559 (LARGE BLOB INSERTS
ARE SLOW AND TROUBLESOME), I found that the sync array that is used
under UNIV_SYNC_DEBUG will store the same level-latch pair over and
over when the latch is recursively locked.  The maximum size of this
sync array was hit when running innodb_bug13450566.test with a 4k
page size because seven different latches were being locked
recursively over and over.  This filled up the available 10,000 slots
in the array and an assert was hit.  When the assert happened, there
were actually only 14 unique latches being tracked for that thread.

This patch add a count field to each slot in this sync array. The
count is > 1 for recursive locking.  This reduces the actual array
size tremendously and can help performance when using UNIV_SYNC_DEBUG
as well as prevent this 'array full' assert when a lot of recursive
locking is happening.  This will require an extra search through the
array to find the previous latch-level slot, but that is offset when
it searches only tens of slots instead of thousands of slots.

Approved by Sunny in http://bur03.no.oracle.com/rb/r/1247/