Bug#14550945 - MTS RECOVERY CRASHES AFTER THE SERVER RESTART WHEN REPOSITORY TYPE WAS CHANGED Bug#14578740 - MTS WITH --RELAY_LOG_INFO_REPOSITORY='FILE' REDUCES WORKERS TO SCREW-UP Bug#14550905 - RECOVERY TABLES FOR MTS DON'T REFLECT CHANGES TO NUMBER OF WORKERS
(Post-itch-and-review changes)

The Bug#14550945 reason is that the replication part of the server bootstrap erroneously 
changed Relay-log info repository type and do not the same to the MTS Worker info.
That made two tightly related piece of info be logically split which led to
a crash when in the following START SLAVE was issued.

Fixed with conducting not letting to migration whenever Worker repositories type is
found to be different than the requested through the new --relay_log_info_repository value.
whether to migrate to another repository type.
Migration is denied and a warning is issued that
the value of --relay_log_info_repository option is altered by the server.
As an extra grace, this patch improves error reporting out of init_slave()
failing branches.
Also a former policy of tolerating error checking of the source or the destination
repositories is changed to conservative. If either of the two found unusable
at checking run in init_slave(), the runtime repositories won't be initialized at all
to prevent running replication slave role on that server.

Bug#14578740 fixes aim at refining and simplifying logics of the file type
repository locating and manipulating.
Fixed with providing necessary context to compose a correct file name
which existence is reliably checked etc.

Bug#14550905 refines post-gaps filling Worker repositories cleanup
when extra Workers info from the last slave session finally becomes redundant.
It is deleted as the extra table records so as extra worker files, unambiguously
now thanks to Bug#14578740 fixes.