Bug#14580874: Limit returns fewer rows than expected with IN + subquery
The query was a semi-join run with materialization strategy, ordering
and limit. When setting up filesort in JOIN::make_tmp_tables_info(),
a query-block-wide limit is applied to a table if there are no other
later tables with conditions or ref accesses. In this case, there was
such table, but the test checked for tab->keyuse, and this field
was not set up in JOIN::setup_materialized_table().