Bug#11750963: "RANGE CHECKED FOR EACH RECORD" IS NOT USED FOR               CONDITIONS WITH OUTER QUERY REFS
When there are range predicates for a table 't1' that depend on 
a non-constant values, the range optimizer is reexecuted after
the table join order has been decided iff 't1' is not the first
table. The range optimizer may then use values read from all
tables 'tx' earlier in the join sequence. This is called 
dynamic range access and is what happens when EXPLAIN shows 
'Range checked for each record'.

However, it is also possible to do dynamic range access for the
first table in a dependent subquery if one of the conditions
for that first table depends on an outer reference. This change
set implements support for using dynamic range access for the 
first table in a dependent subquery.