This is a followup to revno: 4961: 'Update of SPJ component in pre 7.2 branches'.
That commit back ported parts but not all of the online upgrade logic for
SPJ to 7.0. This commit backports the remainder of the online upgarde logic,
that is, the parts concerning the API and the TC blocks.

This commit also removes testcases that will no longer work (because of version
checks) from the daily-basic script. This change should be manually reverted
when merging this commit to 7.2, as the tests are still supposed to work there.
Instead, a new test that check that it is not possible to SPJ API extensions
in pre 7.2 releases have been written (and also added to daily-basic).

Finally, a check of the API version has been added in the API, such that it
is not possible to use SPJ from API client linked with a pre 7.2 API library.
(Before this commit, it would have been possible to run a 7.0 client against
7.2 data nodes.)