BUG#12889121 MYSQLBINLOG CREATES INVALID SQL WHEN USING --VERBOSE OPTION
PROBLEM:                                                    
=========                                                   
When using the mysqlbinlog program with --verbose option    
invalid SQL can be produced when comments appear inside     
BINLOG '...' statements. One way this can happen is when a  
function updates data and this function call is used to do  
a INSERT ... SELECT statement.                              
                                                            
For example: If a statement consists of two events and the  
binlog_format is row based, then the output generated       
by mysqlbinlog tool looks like:                             
##########################################################
  #some header information related to this statement        
  BINLOG '                                                  
    RAW BINARY DATA for the event1                          
    ### verbose section for the event1                      
    RAW BINARY DATA for the events2                         
    ### verbose section for the event2                      
  '/*!*/;                                                   
  ### verbose section for the complete statement            
#########################################################
                                                            
If you take this output and try to execute on mysql client, 
it fails with syntax error because you have invalid         
characters in between the quotes(' ').                      
                                                            
ANALYSIS:                                                   
=========                                                   
The above mentioned problem occurs only if the binlog       
format is set to row or mixed type(In mixed type,           
if the statement is unsafe, we write events in row format). 
                                                            
While printing the log event for the statement(Row format)  
the current existing code is nicely written to take care    
of dividing the header part of the log and body part of it. 
                                                            
Code has two caches to write the log:                       
   IO_CACHE head_cache;                                     
   IO_CACHE body_cache;                                     
                                                            
All the header information will be written to head_cache    
and all the body information will go to body_cache.
The problem comes when --verbose is used. The verbose       
statements are also getting appended to body_cache.         
                                                            
If a statement is having multiple events, these verbose     
statements are coming in between causing the issue.         
                                                            
Fix:
====
 Introduce one more cache for the verbose                   
section(footer_cache). Write all the verbose statements     
in this new cache and at the end of the completion          
after copying head and body cache to the file,              
we can append.                                              
Disadvantage: We are adding one more fwrite call            
                                                           
After the fix the output looks like:                        
======================================                      
#########################################################
  #some header information related to this statement        
  BINLOG '                                                  
    RAW BINARY DATA for the event1                          
    RAW BINARY DATA for the event2                         
  '/*!*/;                                                   
  ### verbose section for the event1                        
  ### verbose section for the event2                        
  ### verbose section for the complete statement            
#########################################################