Bug #14472648 	CONFIGURED DISKCHECKPOINTSPEED EXCEEDED WHEN BACKUPMAXWRITESIZE SET TO HIGH VALU
The DiskCheckpointSpeed mechanism is implemented using 100 millisecond
periods, which each have 1/10th of the configured quota available.

A period is allowed to overflow its quota, with the excess being taken 
from the next period's quota.

However, this overflow was limited to the next period, after that, any
further overflow was ignored.

In cases where large overflows were possible, relative to the 1/10 
DiskCheckPointSpeed quota, this could result in excessive disk writing,
and CPU overhead as a result.

Setting a larger-than standard MaxBackupWriteSize is the primary means
of causing larger-than 2* quota overflows and triggering this bug.

This bug is fixed by using as many subsequent periods as necessary to
'pay off' the quota overflow.

This will result in the data node staying within its quota.

This fix may result in slower LCP in some systems, and reduced CPU usage
during LCP.

A testcase, and an internal DiskCheckPointSpeed verification mechanism
are added to avoid future regressions.