Bug#14117641 SECONDARY INDEX ALREADY CHANGED AFTER ONLINE CREATE INDEX
This is a second try. In the DEBUG_SYNC instrumentation, work around
Bug#14626800 ASSERTION FAILURE IN DEBUG_SYNC().

This bug should only affect very small tables. The bug scenario is as follows.

1. A DML thread has modified the clustered index
   but not yet all secondary indexes.
2. CREATE INDEX picks up the changes directly from the clustered index,
   and completes building the secondary index.
3. The DML thread resumes and finds out that the index that was created
   already contains the change.

Above, DML thread means a thread that is executing an INSERT, UPDATE
or DELETE statement.

A similar bug is not possible in ROLLBACK or purge, because all of
row_undo_ins(), row_undo_mod(), row_purge_del_mark() will modify all
secondary indexes before modifying the clustered index. This means
that the changes would be covered by row_log_online_op() and the
idempotent row_log_apply().

row_ins_sec_index_entry_by_modify(): If the existing record is not
delete-marked, assert that it is a full match of the entry that is to
be inserted (due to INSERT or UPDATE), and assert that the index was
being created online. Then, return DB_SUCCESS, as the record was
already inserted.

row_upd_sec_index_entry(): If the old index record that is to be
delete-marked is not found, do not complain if the index was being
created online. This code is being executed for both UPDATE and
DELETE.

rb:1294 approved by Inaam Rana