Bug#14620839 - THREAD CACHE LEAKS DEBUG INFO ACCROSS SESSIONS
Before this fix, tests such as performance_schema.hostcache*
would pass without the thread cache, and fail with the thread cache.

These tests rely on the SET GLOBAL DEBUG facility.

The thread cache, when reusing an existing thread to run a new session,
polluted the new session debug settings with the old session settings,
causing the failures.

The fix is to reset the debugs settings to a clean initial state,
in the thread cache code, when parking a thread in the cache.