Bug#14576727: Limit returns fewer rows than expected with SELECT as derived table.
The source of this problem is that TABLE_LIST::materializable_is_const()
is calculated based on an estimated_rowcount that is <= 1. Without LIMIT,
JOIN::refine_best_rowcount() enforces that this value is always > 1
for a non-const plan, but with LIMIT, this enforcement does not come into
play.

The problem was in JOIN::refine_best_rowcount().
If a limit was specified, a "best_rowcount" of 1 would not be bumped
to 2, thus the materialized table was considered "const", even though
it was not. Guilhem contributed the fix for this code.