Bug#14001972 ASSERT PAGE_ZIP_GET_SIZE(&BLOCK->PAGE.ZIP) == ZIP_SIZE, ACCESSING SENTINEL
rb://1300
approved by: Marko Makela

In buf_page_get_gen() when ibuf_debug is set there is debug code path
where we try to evict the page if it is found in the buffer pool.
This is to increase chances of ibuf triggering in. In this code path
there is a window where we are not holding any mutex and another thread
can come in and set watch on that page. In that case we choke assuming
that we have found the real page again.