Additional fix for bug#14577463 CONCURRENT DDL & DML LOAD MAY CRASH MYSQLD
The previous fix for this bug was not sufficient, so the bug was reopened.
This fix takes care of another problem related to ref-counting and
freeing of the NDB_SHARE object:

If a NDB_SHARE becomes 'trailing' (It has state==NSS_DROPPED, but there are 
still other active operations which are using the 'to be dropped NDB_SHARE')
then the NDB_SHARE will both be in the 'ndbcluster_dropped_tables'
and 'ndbcluster_open_tables'.

ndbcluster_real_free_share() must therefore always my_hash_delete() a 
'NSS_DROPPED share' from both 'dropped' and 'open' tables before 
it can destruct the NDB_SHARE object.

This fix is relevant for Cluster 7.2 --> only as there are no
'dropped table' hash in 7.0.