Bug#14601802 EXPLAIN DELETE IMPOSSIBLE WHERE, SP,              ASSERTION THD->IS_ERROR() || THD->KILLED
EXLAIN DELETE could cause an assertion on debug builds if
it was used in a stored routine and the explain gave 
"Impossible WHERE" output.

The reason was that the return value from explain_no_table()
was not handled properly, leading the stored routine to
think that EXPLAIN had failed, when it in fact had
succeeded (and therefore not reported any errors).
This was a regression caused by the patch for Bug#11752097.

This patch fixes the problem by assigning the return value
from explain_no_table() to the correct variable.