Bug#14633008 MISLEADING ERROR WHEN TRYING TO UPDATE NON-EXISTING P_S TABLES
Before this change, INSERT / UPDATE / DELETE / LOCK TABLES
performed on unknown performance_schema tables would fail with
errors such as ER_TABLEACCESS_DENIED_ERROR.

For users with the proper grants, the error message for similar cases
with non performance_schema tables is simply ER_NO_SUCH_TABLE.

While this behavior is correct (the performance_schema has some extra security
privileges checks that are built in), it is also confusing.

This fix relaxes the built in privilege checks for the performance schema,
so that DML on tables that do not exist simply fails with ER_NO_SUCH_TABLE.

This is mostly cosmetic, to improve ease of use.