Bug#14621700 SERVER CRASH IN EXECUTING OPTIMIZER_DEBUG_SYNC.TEST
The triggered assert checks that only complete results are
cached in the query cache. It could be triggered if the
query ended with an error and the query cache was turned
on and had a non-zero size.

This problem was introduced by the merge of the patch for
Bug#40264 from 5.0 to 5.1.

This patch fixes the problem by checking if the query has
ended with an error and not caching the results in this case.