Fix for Bug#14103195 CLUSTER WIDE SHUTDOWN ON MULTI-NODE POINTER TOO LARGE IN DBDIH (LINE: 9290)
The SPJ block had no information about which table / indexes which actually
exists, or had been modified or dropped since query execution was started.
Thus, SPJ might submit (DIH-)request for non existing tables or table 
versions, which could crash the DIH block.
      
This fix introduce a simplified dictionary into the SPJ block such that 
SPJ will be able to check the existence / version of a table it is about to 
request an operation on.
      
As this SPJ dictionary has lots in common with the similar dictionary
in TC, the 'Global Dictionary Manager' - GDM module has been created.
TC has then been refactored such that DbtcProxy and DbspjProxy
inherit its 'dictionary proxy parts' from from DbgdmProxy.
            
In order to create & maintain this dictionary, the SPJ block has been 
included in the DICT-loop of which blocks that get dictionary change
notifications.