Fix Bug#14601290 SERVER CRASHED AFTER DESC/CREATE/INSERT/SELECT OPERATIONS OVER RECOVERED TABLE
There is a bug in dict_load_indexes() where if it fails to load some index
due to dict_load_index_id_err then it will just break; out of the loop and
return DB_SUCCESS. If that is the clustered index, then the table object is
left without any indexes which causes a subsequent crash in
row_create_prebuilt() which gets the first index of the table and tries to
de-reference it.

The fix is to return an error at least if we are unable to load the
clustered index in dict_load_indexes().

Approved by:	Jimmy (http://bur03.no.oracle.com/rb/r/1297)