BUG# 12793893 - binlog_innodb timed out in pushbuild.
=== Description ===

The test contained a transaction that ran a loop 2000
times. In every iteration, a row was being being
inserted into a table. 2000 iterations to build up the
long transaction sometimes proved to be too costly on
slow platforms.The purpose behind 2000 iterations was to
make sure that the transaction used the temporary binary
log cache and also exceeded the value of
"binlog_cache_size".

=== Fix ===

I toned down the loop at the same time I made sure it
did not break the purpose.
1) Made the loop run 600 times.
2) Decreased the value of "binlog_cache_size" to 4096,
   the minimum allowed value defined in the manual.