Bug#11755580 ASSERTION IN QUERY CACHE WHEN RUNNING EVENT SCHEDULER Bug#14609893 MYSQLD CRASH OBSERVED WITH "QUERY_CACHE13END_OF_RESULT"
The triggered assert checks that a cached query result is
non-empty at the end of statement execution. It was triggered
if we are executing queries where the result is not sent to the
client. This could e.g. be events (Bug#11755580) or when executing
stored routines containing queries while the server was in 
bootstrap mode (Bug#14609893). The query cache also had to be
turned on and have a non-zero size.

A query result is inserted into the query cache by calls to
Query_cache::insert() which is called from net_write_packet().
However, without active vio, net_write_packet() is skipped.
This meant that the query cache result was never completed,
eventually triggering the assert. It did not happen for normal
statement execution in bootstrap mode since then
Query_cache::end_of_result() is not called. However, stored
routines are executed with the normal code path which contains
calls to Query_cache methods.

This patch fixes the problem by changing Query_cache::store_query()
to not register queries unless there is an active vio (and
Query_cache::insert() therefore subsequentally will be called).
This will cause Query_cache::end_of_result() to not find an 
incomplete result and thereby not trigger the assert.

This bug was only a problem on debug builds. On release builds,
safety code in Query_cache::end_of_result() would simply remove
incomplete query results from the query cache.