BUG#14171756 - mysqlbinlog -v -v does wrong data types.
==== Decription ====

mysqlbinlog -v is used to display a binlog file in verbose
mode, the option -v -v adds comments on column data types.
When data value = NULL, the variable holding data type of
column was not updated. As a result, it used to print data
type of previous column.

==== Fix ====

An if condition was used when data value = NULL. The function
log_event_print_value(), that updated the variable holding
data type of column, was not called inside this if condition.
1) Removed the above mentioned if condition. 
2) Added a check for NULL under log_event_print_value().
3) When data value in first column of a table was NULL, data
   type in mysqlbinlog -v -v was shown as "type = <an integer>".
   Modified this to print data type of column.