Bug#14658739 - PERFORMANCE IMPROVEMENTS IN PFS_LOCK
This fix is a performance improvement to reduce the number
of atomic operations done in the performance schema.

Before this change, a pfs_lock consisted of two attributes,
maintained with atomic operations:
- pfs_lock::m_version
- pfs_lock::m_state

For operations such as:
- pfs_lock::begin_optimistic_lock,
- pfs_lock::end_optimistic_lock,
this caused 2 atomic calls for each method.

First, two consecutive atomic calls are not guaranteed to
be executed atomically (or in the same order), so this
is potentially unsafe.

Secondly, for performance reasons, making only one atomic call
instead of two is better.

With this change, the version number and state attributes
are stored in the same uint32 member, so that both are
changed with only one atomic operation when reading/writing
a pfs_lock from/to memory.

Also, methods such as allocated_to_dirty() used to perform:
- a dirty, non atomic read
- followed by an atomic write.

It is unknown if mixing dirty and atomic operations like this
is valid or not, or if this depends on the CPU architecture used.
To have more robust code, and avoid issues when a value is cached
in a CPU memory cache, these methods now consistently use:
- an atomic read
- an atomic write
to make sure that the code does not use stale data.