BUG#12844279 - VIEWS WITH LOWER() FUNCTION ARE RE-WRITTEN                TO USE LCASE() FUNCTION.
DESCRIPTION:
If a view is created using lower() (string function) for
converting the strings to the lowercase, then select query
is re-written to use the lcase() function.The same problem
existed with the upper() and ucase() functions. According to
MySQL Documentation, lcase() is synonym for lower() function.
lower() and upper() functions are the standard names to be
used. So,lower() should not be replaced by lcase() in query.
lcase() function should be replaced by lower() function in
query definition. Same rule applies for the upper() and
ucase() functions.

ANALYSIS :
Function func_name() in the class Item_func_lcase was
returning "lcase" to the query when called for the function
name. Further, Item_func_lower class was not defined in the
item_create.cc file. Instead, Item_func_lcase was defined.
Same problem existed with the func_name() in Item_func_ucase
class. Item_func_lcase and Item_func_ucase classes are used
when upper() and lower() string functions are invoked.

FIX :
The return value of function func_name in Item_func_lcase and
Item_func_ucase changed to "lower" and "upper" to make queries
adhere to SQL standards. Item_func_lcase and Item_func_ucase
classes renamed to Item_func_lower and Item_func_upper
respectively.