Bug#13742882 - PERFSCHEMA.EVENT_AGGREGATE_NO_U_NO_H AND EVENT_AGGREGATE_NO_A FAIL SPORADICALLY
This fix improves test robustness in the perfschema test suite.

Before this fix, event_aggregate*.test would fail,
with spurious missing statistics for INSERT statements.
Failures would occur only for the last statement of the payload only.

The problem is that the script sends a statement to the server,
and then immediately query performance_schema tables for statement 
statistics.

The reply to the statement to measure might be sent to the client,
but the statement execution is not completed yet in the server.

The fix is to add wait conditions in the test script,
to ensure that the connection that contains the payload is done
with any processing, and blocked in an IDLE wait, waiting for the next 
statement to execute, to make sure the previous statement is completed.