14606694 - ASSERT,RECORD LOCK WAIT IN DICT OPERATION,            DROP TEMP TABLE WHILE IN CREATE INDEX
The assert that is hit in this bug says that if the current transaction
holds dict_operation_lock, it should not have to wait on any blocks in
a SYS_* dictionary table.  But this assert was hit;

	default:
		/* There should never be a lock wait when the
		dictionary latch is reserved in X mode.  Dictionary
		transactions should only acquire locks on dictionary
		tables, not other tables. All access to dictionary
		tables should be covered by dictionary
		transactions. */
		ut_error;

There are two functions that need to assert that dict_operations_lock is
x-owned; dict_insert_tablespace_and_filepath() and dict_update_filepath().
These two functions are called from fil_open_single_table_tablespace()
which is called from 3 places;

1) dict_check_tablespaces_and_store_max_id()
      Run during single threaded startup. But this code does not get
      dict_operations_lock.  That is added in this patch, but since
      this is single threaded, this cannot be the source of the assert
      in this bug.
2) dict_load_table()
      Only had dict_sys->mutex, not dict_operations_lock.  But this call
      to fil_open_single_table_tablespace() sets 'fix_dict = false'
      so it does not call the dictionary functions.
3) row_import_for_mysql().
      This calls row_mysql_lock_data_dictionary(trx); so it has both
      locks already.

The actual way in which this assert was hit was probably that a
transaction created in calls 1 or 3 above needed to be rolled back
for some reason.  If that happens, the UNDO code will not
get dict_operations_lock unless those transactions are identified
as dictionary operations.  This patch does that by starting the
dictionary transaction in these two functions as TRX_DICT_OP_INDEX;

This patch also identifies the sub-transactions in these two functions
as owning an X-lock on dict_operation_lock.

Approved by marko in RB:1311