Bug #14589582 - IMPORT TABLESPACE SHOULD WORK EVEN IF .CFG FILE IS MISSING
WL#5522 forces the requirement that a .cfg (meta-data file) must exist prior
to IMPORT tablespace. This file is required to access the index root pages
and column types. This meta-data is then matched against the importing servers
table meta-data. If there is a mismatch the IMPORT is aborted. This change
allows the IMPORT of a tablespace without the .cfg file. This means that
no meta-data checks are done during import. The index root pages are
determined by scanning the entire .ibd file and checking whether next/prev
pointers are FIL_NULL. Added code to check for free pages. This is done by
caching the keeping the current extent descriptor in memory and then doing
a lookup.

Approved by Jimmy Yang rb://1315