ndb - use 64bit for slack variables in dbacc
in the hash table for primary key there is a slack
variable to keep track on over- and underuse of hash
table memory in words.

the slack variables are 32bit and will not be enough
in future since in theory one can have 2^37 buckets,
each with atleast 28 words. so i makes it 64bit now.