ndb - store cached hash value with element in hash index (dbacc working again)
This patch repairs dbacc and reintroduce the cached hash value disappered in
earlier patch 'removing cached part of hash value for element in hash index'
with revision-id: mauritz.sundell@oracle.com-20120928130216-6lv1u6yxkh5vtkei

this patch reintroduce partial storage of the elements
hash value with the element.
This to not always need to calculate the hash value at
expand or lookup.

With this patch atmost 15 bits of the hash value is stored
for the element.

In linear hashing the lowest bits of the hashvalue is used
as address bits.

As the expand make one more bit of hash value used as address,
one do not need to store that bit, so we shift out one bits
from the stored hash value for each element in both left in 
the splitted bucket and those moved to the new top bucket.

On shrink one need to restore one bit of the stored hashvalue,
and for elements moved from the old top bucket to the new
split bucket a one is shifted in and for the elements already
in the split bucket a zero is shifted in.

Since expand removes one stored hash value bit, fewer bits are
available for quick compare on lookup.  To guarantee for enough
bits after expand one recalculates the full hashvalue if less
than MIN_HASH_COMPARE_BITS bits, and fill up with 15 valid bits
in the stored hash bits.