Bug #14685458 NDB : REPORT BYTES SENT/RECEIVED ACROSS TRANSPORTERS
This patch extends the ndbinfo.transporters table with a number of 
new columns : 
  remote_address
  bytes_sent
  bytes_received

For each transporter (point to point link between
nodes), these columns indicate the address of the 'remote' 
end of the link, the number of bytes sent to that node, and the
number of bytes received from that node.

The byte counts are reset on disconnect.