BUG#14576423 - NO WARNING WHEN THE SERVER IS STARTED                WITH INVALID VALUE FOR QUERY_CACHE_SIZE
DESCRIPTION:
When mysqld is started with an invalid query-cache-size 
parameter (less than 40K according to documentation), query caching
is non-functional, however the global variable "query_cache_size"
still contains the invalid value and no warnings are displayed.

Whereas when query_cache is manually set using
SET GLOBAL command with an invalid value,
a warning is thrown and the value of "query_cache_size"
is reset to 0.


ANALYSIS :
In case of manually setting "query_cache_size" using
SET GLOBAL command, fix_query_cache_size() is fired which is the 
ON_UPDATE callback for Sys_query_cache_size, this callback
checks if resizing the query cache was successful and throws a
warning and resets "query_cache_size" to 0 if an failure occurred. 

However during startup Query Cache is initialized in 
init_server_components() which does not check if Query Cache
initialization failed, hence no warning is thrown and the 
"query_cache_size" global variable is not reset and happens to
contain the same invalid value as passed from the command line.


FIX :
A new function init_server_query_cache() has been introduced
which initializes Query Cache and it also shows a warning and
resets "query_cache_size" to 0 if initialization fails.