Bug#14619935: CRASH IN REGISTER_QUERY_CACHE_TABLE WHEN SELECTING FROM VIEWS
In a set-up where a MERGE VIEW had an underlying TEMPTABLE view,
we asked the (MyISAM-) engine whether we could cache a query.
MyISAM decides this based on whether it thinks the table in question
was changed since the query started. In the given setup, not only is
the object in question a TEMPTABLE, we will also get to it on the 2nd
iteration (of the loop testing all involved objects to see whether any
prevents caching) anyway. We now skip the first test for efficiency.
(In the example that means we don't test the TEMPTABLE of the TEMPTABLE
VIEW while looking at the MERGE VIEW, but when we look at the TEMPTABLE
VIEW. At that point, an existing optimization will fire on account of
the TEMPTABLE VIEW being materializeable.)
This issue would only show up if the sandwiched table was a MyISAM
temptable as then the MyISAM table would try to then try to deref the
table info despite the handler not having been inited yet. For this bug
to show, we'd need a TEMPTABLE (specifically asked for with ALGORITHM,
or forced through use of LIMIT etc.), and it would have to be of MyISAM
type (due do used data type). Lastly, the issue would not show if we
weren't using the query-cache anyway (as in that case, we don't have to
ask the handler whether storing the query is safe).