Fix for Bug#14648712 CALLING PROGERROR WITHOUT THE THIRD ARGUMENT RESULTS IN SIGSEGV
The function ndb_basename() is called from ErrorReporter::handleError() with
'problemData' as argument. As problemData is allowed to be NULL, that used to
crash ndb_basename() when 'strlen()' was called.

This fix will check for NULL argument to ndb_basename() and then
return NULL. The returned NULL value is then later correctly 
handled in ::WriteMessage() and EvenLogger::info() which ::handleErrro()
may then later call with a NULL argument.