Bug#14640599 MEMORY LEAK WHEN EXECUTING STORED ROUTINE EXCEPTION HANDLER
When a SP handler is activated, memory is allocated to hold information
about the condition that caused the activation, the instruction CONTINUE
handlers should return to, etc. Most importantly, memory is allocated to
hold the MESSAGE_TEXT string for the condition.

The problem was that this memory was allocated on the MEM_ROOT belonging
to the stored program. Since this MEM_ROOT is not freed until the
stored program ends, a stored program that causes lots of handler
activations can start using lots of memory. In 5.1 and earlier the
problem did not exist as no condition object was allocated if a condition
was raised with a handler present. However, this behavior lead to
a number of other issues such as Bug#23032.

This patch fixes the problem by allocating this memory on the heap
rather than the SP MEM_ROOT. This means that the memory can be freed
when the handler exits, rather than staying around until the stored
program ends.

This is the 5.6/trunk version of the patch.