Fix Bug#14658648 XA ROLLBACK (DISTRIBUTED DATABASE) NOT WORKING WITH QUERY CACHE ENABLED
The bug is caused by the fact that we do not restore table level locks
for XA PREPARED transactions during crash recovery. Doing so may require
changing the on-disk format, so we choose to disable query cache completely
for all tables if XA PREPARED and recovered transactions exist in the
system, even though they may not have touched the table in question in
row_search_check_if_query_cache_permitted().

We introduce a new trx_sys->n_prepared_recovered_trx member that keeps count
of the number of recovered transactions that are in prepared state, so we
can quickly determine in row_search_check_if_query_cache_permitted() whether
the query cache should be disabled.

Approved by:	Marko (rb:1351)