Fix for Bug#14604223 - INCORRECT RESULTS FOR DESCENDING RANGE SCAN WITH                        INDEX CONDITION PUSHDOWN
When a descending range scan uses index condition pushdown the storage
engine needs to know about the end of the range in order to avoid that
the storage engine reads to the beginning of the index. This was
implemented as a fix to Bug#14503142. The code in
QUICK_SELECT_DESC::get_next() uses both ha_index_prev() and
ha_index_next_same() to read data. Setting the end range is only
needed when using ha_index_prev(). The wrong result in this bug
happens when the range scan consists of multiple ranges. If the first
range sets the end range, it is necessary to either set the end range
for all following ranges or to let the storage engine know that there
is no end range. The later part was not implemented. If the second
range to read was either an EQ_RANGE or did not have an end range, we
failed to let the storage engine know this. Due to this, the storage
engine would continue to evaluate records against the end range from
the first range. This caused no records to be returned when reading
from the second range.

The fix for this problem is to call set_end_range(NULL,...) to let the
storage engine know that the next range does not have an end range. This
should be done both when the range is an EQ_RANGE and when the range does
not have a lower end range.