Bug#14630263 CREATE UNIQUE INDEX CREATES CORRUPTED INDEX ON NULLABLE COLUMN (second patch)
row_log_apply_op_low(): When the key fields match on a unique index
but some of them are NULL, do not replace a partially matching
preceding record, but insert the new record after it.

This fixes a case that was forgotten in the previous push (rb:1314).

rb:1361 approved by Jimmy Yang